+++ test/time_test.dart
@@ -9,31 +9,31 @@ void main() {
 
   test('Time', () async {
 
-    var now = DateTime.now().toUtc();
+    final now = DateTime.now().toUtc();
 
     int value = 0;
 
-    var s = TimeScheduler();
-    var t1 = s.run(()=>(value=1), now.add(Duration(seconds: 1)));
-    var t2 = s.run(()=>(value=2), now.add(Duration(seconds: 2)));
-    var t3 = s.run(()=>(value=3), now.add(Duration(seconds: 3)));
-    var t4 = s.run(()=>(value=4), now.add(Duration(seconds: 4)));
+    final s = TimeScheduler();
+    s.run(()=>(value=1), now.add(Duration(seconds: 1)));
+    s.run(()=>(value=2), now.add(Duration(seconds: 2)));
+    final t3 = s.run(()=>(value=3), now.add(Duration(seconds: 3)));
+    s.run(()=>(value=4), now.add(Duration(seconds: 4)));
 
     expect(value, 0);
-    await Future.delayed(Duration(milliseconds: 1100));
+    await Future<void>.delayed(Duration(milliseconds: 1100));
     expect(value, 1);
-    await Future.delayed(Duration(milliseconds: 1100));
+    await Future<void>.delayed(Duration(milliseconds: 1100));
     expect(value, 2);
     t3.willRun = false;
-    await Future.delayed(Duration(milliseconds: 1100));
+    await Future<void>.delayed(Duration(milliseconds: 1100));
     expect(value, 2);
     s.dispose();
-    await Future.delayed(Duration(milliseconds: 1100));
+    await Future<void>.delayed(Duration(milliseconds: 1100));
     expect(value, 2);
   });
 
   test('Cannot add to disposed', () async {
-    var s = TimeScheduler();
+    final s = TimeScheduler();
     s.run(() => {}, DateTime.now().add(Duration(seconds: 1)));
     s.dispose();
     expect(()=>s.run(() => {}, DateTime.now().add(Duration(seconds: 1))), throwsStateError);