+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Result of `GetterFunc` changed to `FutureOr`
+
 # 0.3.1
 
 - Added TimeScheduler